Change type Volume property to Decimal in Candle-class. WARNING! This can be a breaking change in a number of cases

+++ src/QuikSharp/DataStructures/Candle.cs
@@ -38,7 +38,7 @@ namespace QuikSharp.DataStructures
         /// Объем последней сделки
         /// </summary>
         [JsonProperty("volume")]
-        public int Volume { get; set; }
+        public decimal Volume { get; set; }
 
         /// <summary>
         /// Дата и время
